requirements.txt
langchain
python-dotenv
openai

torch
numpy
pandas
chardet
aiohttp
charset-normalizer==3.1.0

google-search-results
